+++ ReleaseNotes.md
@@ -12,7 +12,7 @@
 - Update of French verison (by Lanneval and Nono)
 - Some redesign of the UI to have a more unified medieval flavor (by @Sylph)
 - Rearranged languague mnemonics of common usage into a generic category (the older ones will be keep for a few versions to allow translators to catch up)
-- New active effects for source quality modifiers of seasonal activities
+- New active effects for source quality modifiers of seasonal activities (reading and writting have no effect yet)
 
 ### Bug fixes